requirements.txt
Flask==2.1.2
pathlib2==2.3.7
tinydb==4.7.0
pyyaml==6.0
markdown==3.3.7